Create Branch using if statement.cs
One of the common mathematical problems is to find out which quadrant a given point belongs to. Quadrants have numbers from 1 to 4, as shown in the figure below. "Quadrantn" means "the nth quadrant".

For example, A, whose coordinates are (12, 5) belongs to the first quadrant because both x and y coordinates are positive. Point B belongs to the second quadrant because the x-coordinate is negative and the y-coordinate is positive.

Write a program that takes the coordinates of the points and determines which quadrant they belong to. However, assume that both the x and y coordinates are positive or negative.