Create alarmclock.cs
Full-time wakes up to the alarm every morning. It would be fortunate if I woke up right after hearing the alarm, but I am always late for school because I want to sleep a little longer.

Full-time tried everything, but his desire to sleep a little longer could not get rid of anything.

Chang-young, who felt sorry for such a full-time employee, recommended a way to use it.

It's "Set the alarm 45 minutes earlier."

This method is simple. It is to change the alarm to a time 45 minutes ahead of the original setting. This is because if I hear the alarm anyway, I will turn off the alarm and sleep a little more. Using this method, you can feel that you slept more every morning, and you are not late for school.

If Chang-young's method is used when given the alarm time set by the current full-time employee, write a program to find out when to fix it.

Input
The first line is given two integers H and M. (0 H H 23 23, 0 M M 5 59), and this means the alarm time H time M set by the current full-time employee.

The input time uses a 24-hour representation. In the 24-hour expression, the start of the day is 0:0 (midnight), and the end is 23:59 (one minute before midnight the next day). When representing time, unnecessary zero is not used.

Output
When full-time uses Chang-young's method in the first line, output the alarm time to be set. (You can print it out in the same form as the input.)